+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to similar are documented here.
 
 * Added grapheme and character level diffing utilities.
 * `DiffOp::as_tag_tuple` is now taking the argument by reference.
+* Added `TextDiff::ratio`.
 
 ## 0.2.0